+++ src/main/java/com/deliveryfood/controller/OrderController.java
@@ -1,4 +1,4 @@
-package com.deliveryfood.service;
+package com.deliveryfood.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
